move internal Date extension into its own file
