+++ DVB/Classes/DataTypes/Stop.swift
@@ -22,23 +22,6 @@ public struct FindResponse {
     }
 }
 
-// TODO: Put this somewhere else
-internal extension Date {
-    // Init with a string of the format "/Date(1487458060455+0100)/"
-    init?(from dateString: String) {
-        let components = dateString
-            .replacingOccurrences(of: "/Date(", with: "")
-            .replacingOccurrences(of: ")/", with: "")
-            .components(separatedBy: "+") // FIXME: This will surely break...
-
-        guard let millis = Int(components[0]) else { return nil }
-        guard let tz = Int(components[1]) else { return nil }
-
-        let seconds = Double(millis) / 1000 + Double(tz * 60 * 60)
-        self = Date(timeIntervalSince1970: seconds)
-    }
-}
-
 /// A place where a bus, tram or whatever can stop.
 public struct Stop {
     public let id: String
